Handle uncontrolled ordering of static object lifetimes
Construct the registered log handlers on first use with a local static
function. Maintain a weak reference to this in every Handler so that they
can be unregistered safely when they are destroyed.

It is still unsafe to call logging functions at exit because the local
static registered log handlers may have been destroyed.